Add NoMsg argument, change how msg input is handled
Before this, message input was considered to be used if at least one Write call has been made. However, io.Copy doesn't even call Write if message length is zero, so our implementation considered it as an absent message block (like in stream cipher) rather than zero-length message block (like in hash and MAC), which is treated differently in Skein.

This change adds NoMsg field to Args. If it's false (default), it indicates that the message input is used (thus absent message is equal to zero-length message). If it's true, the message input is not used, and calling Write will return an error.

NewStream specifies NoMsg as true, while all other primitives use false.

The change breaks compatibility with custom constructions based on Hash that assumed that not calling Write gives the same result as currently NoMsg. It will also produce correct, but different hashes for zero-length messages, when used with io.Copy or without calling Write.

For detailed explanation, see: https://github.com/dchest/skein/pull/1#issuecomment-343725160

Thanks to @martinlindhe for spotting the issue and writing the test case.